move `findFreePort()` out of `index.ts`
