+++ src/docker/ports.ts
@@ -1,0 +1,14 @@
+import * as net from "net";
+
+/** Look for an available port on the host machine and return it. */
+export async function findFreePort(): Promise<number> {
+  return new Promise((resolve, reject) => {
+    const server = net.createServer();
+    server.unref();
+    server.on("error", reject);
+    server.listen(0, () => {
+      const port = (server.address() as net.AddressInfo).port;
+      server.close(() => resolve(port));
+    });
+  });
+}
